+++ tests/NodaMoney.Tests/MoneyCalculationsSpec.cs
@@ -23,12 +23,14 @@ namespace NodaMoney.Tests.MoneyCalculationsSpec
         //    // Money result = subject.Perform(fx);
         //}
 
-        //[Fact]
-        //public void When1_Then()
-        //{
-        //    var x = default(Money);
-        //    var c = default(Currency);
-        //    Console.WriteLine();
-        //}
+        [Fact]
+        public void When1_Then()
+        {
+            Currency x = default;
+            Money c = default;
+
+            //Console.WriteLine(c);
+            //Console.WriteLine(x);
+        }
     }
 }
