+++ tests/NodaMoney.Tests/Serialization/JavaScriptSerializableSpec.cs
@@ -3,64 +3,63 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Web.Script.Serialization;
 using FluentAssertions;
+using NodaMoney.Serialization.AspNet;
 using Xunit;
 
-namespace NodaMoney.Serialization.AspNet.Tests
+namespace NodaMoney.Tests.Serialization.AspNet
 {
-    public static class JavaScriptSerializableSpec
+    public class GivenIWantToSerializeMoneyWithJavaScriptSerializer
     {
-        public class GivenIWantToSerializeMoneyWithJavaScriptSerializer
+        public static IEnumerable<object[]> TestData => new[]
         {
-            public static IEnumerable<object[]> TestData => new[]
-            {
                 new object[] { new Money(765m, Currency.FromCode("JPY")) },
                 new object[] { new Money(765.43m, Currency.FromCode("EUR")) }
             };
 
-            [Theory]
-            [MemberData(nameof(TestData))]
-            public void WhenSerializing_ThenThisShouldSucceed(Money money)
-            {
-                var jsSerializer = new JavaScriptSerializer();
-                jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
-
-                var json = jsSerializer.Serialize(money);
-                // Console.WriteLine(json);
-                var clone = jsSerializer.Deserialize<Money>(json);
-
-                clone.Should().Be(money);
-            }
-        }
-
-        public class GivenIWantToDeserializeMoneyWithJavaScriptSerializer
+        [Theory]
+        [MemberData(nameof(TestData))]
+        public void WhenSerializing_ThenThisShouldSucceed(Money money)
         {
-            private static string CurrentCultureCode = new RegionInfo(CultureInfo.CurrentCulture.LCID).ISOCurrencySymbol;
+            var jsSerializer = new JavaScriptSerializer();
+            jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
 
-            public static IEnumerable<object[]> ValidJsonData => new[]
-            {
+            var json = jsSerializer.Serialize(money);
+            // Console.WriteLine(json);
+            var clone = jsSerializer.Deserialize<Money>(json);
+
+            clone.Should().Be(money);
+        }
+    }
+
+    public class GivenIWantToDeserializeMoneyWithJavaScriptSerializer
+    {
+        private static string CurrentCultureCode = new RegionInfo(CultureInfo.CurrentCulture.LCID).ISOCurrencySymbol;
+
+        public static IEnumerable<object[]> ValidJsonData => new[]
+        {
                 new object[] { $"{{ amount: '200', currency: '{CurrentCultureCode}' }}", },
                 new object[] { $"{{ amount: 200, currency: '{CurrentCultureCode}' }}" },
                 new object[] { $"{{ currency: '{CurrentCultureCode}', amount: 200 }}" },
                 new object[] { $"{{ currency: '{CurrentCultureCode}', amount: '200' }}" }
             };
 
-            public static IEnumerable<object[]> InvalidJsonData => new[]
-            {
+        public static IEnumerable<object[]> InvalidJsonData => new[]
+        {
                 new object[] { "{ amount: '200' }" },
                 new object[] { "{ amount: 200 }" },
                 new object[] { $"{{ currency: '{CurrentCultureCode}' }}" },
             };
 
-            public static IEnumerable<object[]> ValidNestedJsonData => new[]
-            {
+        public static IEnumerable<object[]> ValidNestedJsonData => new[]
+        {
                 new object[] { $"{{ cash: {{ amount: '200', currency: '{CurrentCultureCode}' }} }}", },
                 new object[] { $"{{ cash: {{ amount: 200, currency: '{CurrentCultureCode}' }} }}" },
                 new object[] { $"{{ cash: {{ currency: '{CurrentCultureCode}', amount: 200 }} }}" },
                 new object[] { $"{{ cash: {{ currency: '{CurrentCultureCode}', amount: '200' }} }}" }
             };
 
-            public static IEnumerable<object[]> ValidNestedNullableJsonData => new[]
-            {
+        public static IEnumerable<object[]> ValidNestedNullableJsonData => new[]
+        {
                 new object[] { $"{{ cash: {{ amount: '200', currency: '{CurrentCultureCode}' }} }}", },
                 new object[] { $"{{ cash: {{ amount: 200, currency: '{CurrentCultureCode}' }} }}" },
                 new object[] { $"{{ cash: {{ currency: '{CurrentCultureCode}', amount: 200 }} }}" },
@@ -68,79 +67,78 @@ namespace NodaMoney.Serialization.AspNet.Tests
                 new object[] { $"{{ cash: null }}" },
             };
 
-            [Theory]
-            [MemberData(nameof(ValidJsonData))]
-            public void WhenDeserializingWithValidJSON_ThenThisShouldSucceed(string json)
-            {
-                var money = new Money(200, Currency.FromCode(CurrentCultureCode));
+        [Theory]
+        [MemberData(nameof(ValidJsonData))]
+        public void WhenDeserializingWithValidJSON_ThenThisShouldSucceed(string json)
+        {
+            var money = new Money(200, Currency.FromCode(CurrentCultureCode));
 
-                var jsSerializer = new JavaScriptSerializer();
-                jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
+            var jsSerializer = new JavaScriptSerializer();
+            jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
 
-                // Console.WriteLine(json);
-                var clone = jsSerializer.Deserialize<Money>(json);
+            // Console.WriteLine(json);
+            var clone = jsSerializer.Deserialize<Money>(json);
 
-                clone.Should().Be(money);
-            }
+            clone.Should().Be(money);
+        }
 
-            [Theory]
-            [MemberData(nameof(InvalidJsonData))]
-            public void WhenDeserializingWithInvalidJSON_ThenThisShouldFail(string json)
-            {
-                var money = new Money(200, Currency.FromCode(CurrentCultureCode));
+        [Theory]
+        [MemberData(nameof(InvalidJsonData))]
+        public void WhenDeserializingWithInvalidJSON_ThenThisShouldFail(string json)
+        {
+            var money = new Money(200, Currency.FromCode(CurrentCultureCode));
 
-                var jsSerializer = new JavaScriptSerializer();
-                jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
+            var jsSerializer = new JavaScriptSerializer();
+            jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
 
-                var exception = Record.Exception(() =>
-                    jsSerializer.Deserialize<Money>(json)
-                );
+            var exception = Record.Exception(() =>
+                jsSerializer.Deserialize<Money>(json)
+            );
 
-                exception.Should().BeOfType<ArgumentNullException>();
-            }
+            exception.Should().BeOfType<ArgumentNullException>();
+        }
 
-            private class TypeWithMoneyProperty
-            {
-                public Money Cash { get; set; }
-            }
+        private class TypeWithMoneyProperty
+        {
+            public Money Cash { get; set; }
+        }
 
-            [Theory]
-            [MemberData(nameof(ValidNestedJsonData))]
-            public void WhenDeserializingWithNested_ThenThisShouldSucceed(string json)
-            {
-                var money = new Money(200, Currency.FromCode(CurrentCultureCode));
+        [Theory]
+        [MemberData(nameof(ValidNestedJsonData))]
+        public void WhenDeserializingWithNested_ThenThisShouldSucceed(string json)
+        {
+            var money = new Money(200, Currency.FromCode(CurrentCultureCode));
 
-                var jsSerializer = new JavaScriptSerializer();
-                jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
+            var jsSerializer = new JavaScriptSerializer();
+            jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
 
-                // Console.WriteLine(json);
-                var clone = jsSerializer.Deserialize<TypeWithMoneyProperty>(json);
+            // Console.WriteLine(json);
+            var clone = jsSerializer.Deserialize<TypeWithMoneyProperty>(json);
 
+            clone.Cash.Should().Be(money);
+        }
+
+        private class TypeWithNullableMoneyProperty
+        {
+            public Money? Cash { get; set; }
+        }
+
+        [Theory]
+        [MemberData(nameof(ValidNestedNullableJsonData))]
+        public void WhenDeserializingWithNestedNullable_ThenThisShouldSucceed(string json)
+        {
+            var money = new Money(200, Currency.FromCode(CurrentCultureCode));
+
+            var jsSerializer = new JavaScriptSerializer();
+            jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
+
+            // Console.WriteLine(json);
+            var clone = jsSerializer.Deserialize<TypeWithNullableMoneyProperty>(json);
+
+            if (!json.Contains("null"))
                 clone.Cash.Should().Be(money);
-            }
-
-            private class TypeWithNullableMoneyProperty
-            {
-                public Money? Cash { get; set; }
-            }
-
-            [Theory]
-            [MemberData(nameof(ValidNestedNullableJsonData))]
-            public void WhenDeserializingWithNestedNullable_ThenThisShouldSucceed(string json)
-            {
-                var money = new Money(200, Currency.FromCode(CurrentCultureCode));
-
-                var jsSerializer = new JavaScriptSerializer();
-                jsSerializer.RegisterConverters(new JavaScriptConverter[] { new MoneyJavaScriptConverter() });
-
-                // Console.WriteLine(json);
-                var clone = jsSerializer.Deserialize<TypeWithNullableMoneyProperty>(json);
-
-                if (!json.Contains("null"))
-                    clone.Cash.Should().Be(money);
-                else
-                    clone.Cash.Should().BeNull();
-            }
+            else
+                clone.Cash.Should().BeNull();
         }
     }
 }
