+++ tests/NodaMoney.Tests/Serialization/RavenDbSerializationSpec.cs
@@ -7,7 +7,7 @@ namespace NodaMoney.Tests.Serialization.RavenDbSerializationSpec
 {
     public class GivenIWantToStoreInRavenDb : RavenTestDriver
     {
-        [Fact] //(Skip = "Microsoft.NETCore.App 2.1.10 not installed on buildserver")]
+        [Fact] //(Skip = "Microsoft.NETCore.App 2.1.10 not installed on build server")]
         public void WhenMoneyAsRoot_ThenThisMustWork()
         {
             Money euros = new Money(123.56, "EUR");
@@ -34,7 +34,7 @@ namespace NodaMoney.Tests.Serialization.RavenDbSerializationSpec
             }
         }
 
-        [Fact] //(Skip = "Microsoft.NETCore.App 2.1.10 not installed on buildserver")]
+        [Fact] //(Skip = "Microsoft.NETCore.App 2.1.10 not installed on build server")]
         public void WhenObjectWithMoneyAttribute_ThenThisMustWork()
         {
             SampleData sample = new SampleData { Name = "Test", Price = new Money(123.56, "EUR"), BaseCurrency = Currency.FromCode("USD") };
